feat(useContractValidation): implement contract validation on services
