+++ src/hooks/useDeleteRequest.ts
@@ -3,6 +3,7 @@ import { useNavigate, useLocation } from "react-router-dom";
 
 import { useErrorFlag } from "@hooks/useErrorFlag";
 import { useHeaders } from "@hooks/useHeaders";
+import { useContractValidation } from "@hooks/useContractValidation";
 import { deleteHumanResourceRequest } from "@services/humanResourcesRequest/deleteHumanResourceRequest";
 
 export function useDeleteRequest<T extends { requestId?: string }>(
@@ -13,6 +14,8 @@ export function useDeleteRequest<T extends { requestId?: string }>(
   const [isDeleting, setIsDeleting] = useState(false);
   const [showFlag, setShowFlag] = useState(false);
   const { getHeaders } = useHeaders();
+
+  useContractValidation();
 
   useErrorFlag(
     showFlag,
